minimap working better but not done yet

+++ src/parsing/starting_position.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   starting_position.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jhesso <jhesso@student.hive.fi>            +#+  +:+       +#+        */
+/*   By: dgerguri <dgerguri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/11 03:36:06 by jhesso            #+#    #+#             */
-/*   Updated: 2023/11/11 03:45:05 by jhesso           ###   ########.fr       */
+/*   Updated: 2023/11/11 16:05:45 by dgerguri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ static void	get_angle(t_map_data *data, float angle)
 
 void	starting_point(t_map_data *d, t_vector pos, bool *p_start, char c)
 {
-	d->pos.x = pos.x * SIZE_B + ((SIZE_B - SIZE_P) / 2);
-	d->pos.y = pos.y * SIZE_B + ((SIZE_B - SIZE_P) / 2);
+	d->pos.x = pos.x * (float)SIZE_B + (((float)SIZE_B - (float)SIZE_P) / 2);
+	d->pos.y = pos.y * (float)SIZE_B + (((float)SIZE_B - (float)SIZE_P) / 2);
 	*(p_start) = true;
 	if (c == 'N')
 		get_angle(d, 90.0f);
